refactor(search applications): rename engine api integration file
