+++ x-pack/plugins/enterprise_search/public/applications/applications/components/search_application/connect/search_application_api.tsx
@@ -29,8 +29,8 @@ import { KibanaLogic } from '../../../../shared/kibana';
 
 import { SearchApplicationViewLogic } from '../search_application_view_logic';
 
-import { EngineApiIntegrationStage } from './engine_api_integration';
 import { GenerateEngineApiKeyModal } from './generate_engine_api_key_modal/generate_engine_api_key_modal';
+import { EngineApiIntegrationStage } from './search_application_api_integration';
 import { SearchApplicationApiLogic } from './search_application_api_logic';
 
 export const elasticsearchUrl = (cloudContext: CloudDetails): string => {
